requirements.txt
python-telegram-bot[job-queue]
yargy
APScheduler
python-dotenv
openai
aiohttp>=3.8.0